pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2markdown"
version = "0.1.0"
description = "High-fidelity PDF to Markdown converter with OCR support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "PDF2MD Team"}
]
keywords = ["pdf", "markdown", "converter", "ocr", "tables"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core conversion libraries
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.7",

    # Image processing
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",

    # OCR support
    "pytesseract>=0.3.10",

    # CLI interface
    "typer[all]>=0.9.0",
    "rich>=13.0.0",

    # Configuration
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Utilities
    "python-magic>=0.4.27",
    "pathvalidate>=3.0.0",
]

[project.optional-dependencies]
# High-accuracy converter (optional, large download)
accurate = [
    "marker-pdf>=0.2.0",
    "torch>=2.0.0",
]

# Web interfaces
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.28.0",
]

# Async processing
async = [
    "celery>=5.3.0",
    "redis>=5.0.0",
    "flower>=2.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

# All optional dependencies
all = [
    "pdf2markdown[accurate,web,async,dev]"
]

[project.scripts]
pdf2md = "pdf2markdown.cli.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/pdf2markdown"
Documentation = "https://github.com/yourusername/pdf2markdown#readme"
Repository = "https://github.com/yourusername/pdf2markdown"
Issues = "https://github.com/yourusername/pdf2markdown/issues"

[tool.setuptools]
packages = ["pdf2markdown"]

[tool.setuptools.package-dir]
pdf2markdown = "src"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
